+++ src/inspect_ai/_view/www/src/app/samples/list/SampleRow.tsx
@@ -4,6 +4,7 @@ import { SampleSummary } from "../../../client/api/types";
 import { PulsingDots } from "../../../components/PulsingDots";
 import { useStore } from "../../../state/store";
 import { arrayToString, inputString } from "../../../utils/format";
+import { truncateMarkdown } from "../../../utils/markdown";
 import { isVscode } from "../../../utils/vscode";
 import { RenderedText } from "../../content/RenderedText";
 import { SampleErrorView } from "../error/SampleErrorView";
@@ -21,8 +22,6 @@ interface SampleRowProps {
   showSample: () => void;
   sampleUrl?: string;
 }
-
-const kMaxRowTextSize = 1024 * 5;
 
 export const SampleRow: FC<SampleRowProps> = ({
   id,
@@ -95,9 +94,10 @@ export const SampleRow: FC<SampleRowProps> = ({
       >
         {!showingSampleDialog ? (
           <RenderedText
-            markdown={inputString(sample.input)
-              .join(" ")
-              .slice(0, kMaxRowTextSize)}
+            markdown={truncateMarkdown(
+              inputString(sample.input).join(" "),
+              250,
+            )}
             forceRender={true}
             omitMedia={true}
           />
@@ -106,7 +106,7 @@ export const SampleRow: FC<SampleRowProps> = ({
       <div className={clsx("sample-target", "three-line-clamp", styles.cell)}>
         {sample?.target && !showingSampleDialog ? (
           <RenderedText
-            markdown={arrayToString(sample.target).slice(0, kMaxRowTextSize)}
+            markdown={truncateMarkdown(arrayToString(sample.target), 250)}
             className={clsx("no-last-para-padding", styles.noLeft)}
             forceRender={true}
             omitMedia={true}
@@ -116,7 +116,7 @@ export const SampleRow: FC<SampleRowProps> = ({
       <div className={clsx("sample-answer", "three-line-clamp", styles.cell)}>
         {sample && !showingSampleDialog ? (
           <RenderedText
-            markdown={(answer || "").slice(0, kMaxRowTextSize)}
+            markdown={truncateMarkdown(answer || "", 250)}
             className={clsx("no-last-para-padding", styles.noLeft)}
             forceRender={true}
             omitMedia={true}