+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Analysis: Read all samples at once in implementation of `samples_df()`.
+- Inspect View: Improve sample summary truncation (use markdown truncation instead of line clamping).
 
 ## 0.3.162 (18 January 2026)